fix: separate m365 token provider wrapper with m365 interactive login to avoid dependency issue
