+++ packages/tests/.mocharc.js
@@ -2,5 +2,5 @@ module.exports = {
     timeout: 20 * 60 * 1000,
     exit: true,
     reporter: "mochawesome",
-    require: ["ts-node/register", "./src/e2e/setup.ts"]
+    require: ["ts-node/register"]
 };
